+++ cmake/onnxruntime_python.cmake
@@ -173,11 +173,14 @@ if (onnxruntime_ENABLE_TRAINING)
   file(GLOB onnxruntime_python_capi_training_srcs CONFIGURE_DEPENDS
     "${ORTTRAINING_SOURCE_DIR}/python/training/*.py"
   )
-  file(GLOB onnxruntime_python_capi_training_amp_srcs CONFIGURE_DEPENDS
-    "${ORTTRAINING_SOURCE_DIR}/python/training/amp/*.py"
+  file(GLOB onnxruntime_python_root_srcs CONFIGURE_DEPENDS
+    "${ORTTRAINING_SOURCE_DIR}/python/experimental/*.py"
   )
-  file(GLOB onnxruntime_python_capi_training_optim_srcs CONFIGURE_DEPENDS
-    "${ORTTRAINING_SOURCE_DIR}/python/training/optim/*.py"
+  file(GLOB onnxruntime_python_amp_srcs CONFIGURE_DEPENDS
+    "${ORTTRAINING_SOURCE_DIR}/python/experimental/amp/*.py"
+  )
+  file(GLOB onnxruntime_python_optim_srcs CONFIGURE_DEPENDS
+    "${ORTTRAINING_SOURCE_DIR}/python/experimental/optim/*.py"
   )
 else()
   file(GLOB onnxruntime_python_capi_training_srcs CONFIGURE_DEPENDS
@@ -215,8 +218,9 @@ add_custom_command(
   COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/backend
   COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi
   COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/training
-  COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/training/amp
-  COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/training/optim
+  COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/experimental
+  COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/experimental/amp
+  COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/experimental/optim
   COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/datasets
   COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/tools
   COMMAND ${CMAKE_COMMAND} -E make_directory $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/tools/featurizer_ops
@@ -246,11 +250,14 @@ add_custom_command(
       ${onnxruntime_python_capi_training_srcs}
       $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/training/
   COMMAND ${CMAKE_COMMAND} -E copy
-      ${onnxruntime_python_capi_training_amp_srcs}
-      $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/training/amp/
+      ${onnxruntime_python_root_srcs}
+      $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/experimental/
   COMMAND ${CMAKE_COMMAND} -E copy
-      ${onnxruntime_python_capi_training_optim_srcs}
-      $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/training/optim/
+      ${onnxruntime_python_amp_srcs}
+      $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/experimental/amp/
+  COMMAND ${CMAKE_COMMAND} -E copy
+      ${onnxruntime_python_optim_srcs}
+      $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/experimental/optim/
   COMMAND ${CMAKE_COMMAND} -E copy
       $<TARGET_FILE:onnxruntime_pybind11_state>
       $<TARGET_FILE_DIR:${test_data_target}>/onnxruntime/capi/
