+++ orttraining/orttraining/python/experimental/__init__.py
@@ -4,3 +4,7 @@
 # --------------------------------------------------------------------------
 from onnxruntime.capi._pybind_state import TrainingParameters
 from onnxruntime.capi.training.training_session import TrainingSession
+
+from .orttrainer_options import ORTTrainerOptions
+from .orttrainer import ORTTrainer, TrainStepInfo
+from . import amp, optim, model_desc_validation
