+++ orttraining/orttraining/test/python/_test_helpers.py
@@ -4,7 +4,7 @@ import sys
 import torch
 
 from numpy.testing import assert_allclose
-from onnxruntime.capi.training import orttrainer
+from onnxruntime.experimental import orttrainer
 from onnxruntime.capi.ort_trainer import ORTTrainer as Legacy_ORTTrainer
