+++ setup.py
@@ -227,6 +227,9 @@ packages = [
     'onnxruntime.backend',
     'onnxruntime.capi',
     'onnxruntime.capi.training',
+    'onnxruntime.experimental',
+    'onnxruntime.experimental.amp',
+    'onnxruntime.experimental.optim',
     'onnxruntime.datasets',
     'onnxruntime.tools',
     'onnxruntime.quantization',
